Restructured React files
