+++ src/main.jsx
@@ -1,11 +1,11 @@
-import './index.css'
-import App from './App.jsx'
-import ProductPage from './ProductPage.jsx'
-import ShopAll from './shopAll.jsx'
+import "./index.css";
+import App from "./App.jsx";
+import ProductPage from "./products/ProductPage.jsx";
+import ShopAll from "./shopAll/shopAll.jsx";
 import React from "react";
 import ReactDOM from "react-dom/client";
-import { createBrowserRouter, RouterProvider } from 'react-router-dom';
-import { ShoppingCart } from './shoppingCart.jsx';
+import { createBrowserRouter, RouterProvider } from "react-router-dom";
+import { ShoppingCart } from "./shoppingCart/shoppingCart.jsx";
 
 const root = ReactDOM.createRoot(document.getElementById("root"));
 
@@ -13,7 +13,7 @@ const router = createBrowserRouter([
   { path: "/", element: <App /> },
   { path: "/productPage/:id", element: <ProductPage /> },
   { path: "/shopAll/:category", element: <ShopAll /> },
-  { path: "/cart", element: <ShoppingCart />}
+  { path: "/cart", element: <ShoppingCart /> },
 
   /*,
   { path: "/products", element: <Products /> },
@@ -24,6 +24,6 @@ const router = createBrowserRouter([
 
 root.render(
   //<React.StrictMode>
-    <RouterProvider router={router} />
+  <RouterProvider router={router} />
   //</React.StrictMode>
-); 
+);
