+++ src/shopAll/shopAll.css
